refactor(memory_sync): internalize the cap math into the orchestrator
With all six providers now riding `run_sync`, the per-sync `max_items`
cap math is consumed only by the orchestrator. Relocate `ItemCap` and
`pages_for_max_items` (plus their unit tests) out of the shared
`providers::helpers` module and into `orchestrator.rs` as private items,
so the cap/window/dedup rule lives in exactly one place — the payoff of
#3333.

`epoch_floor_from_depth` stays in `helpers` (it is a provider-facing
window helper: `gmail::source` builds an `after:<epoch>` filter from it,
not orchestrator-internal cap math). Mechanical move — no behavior
change; all memory_sync provider/unit/e2e tests stay green.

Closes #3902
